+++ Vector2D.hpp
@@ -6,11 +6,11 @@
 class Vector2D
 {
 public:
-    float x;
-    float y;
+    double x;
+    double y;
 
     Vector2D();
-    Vector2D(float x, float y);
+    Vector2D(double x, double y);
 
     Vector2D& Add(const Vector2D& vec);
     Vector2D& Subtract(const Vector2D& vec);
